feat(devenv): Add support for modules, automate and refactor build process
