+++ .gitignore
@@ -7,6 +7,9 @@ target/
 # More information here https://doc.rust-lang.org/cargo/guide/cargo-toml-vs-cargo-lock.html
 Cargo.lock
 
+# Generated rust-analyzer file
+rust-project.json
+
 # These are backup files generated by rustfmt
 **/*.rs.bk
 
@@ -14,6 +17,5 @@ Cargo.lock
 **/*.[o,mod,ko]*
 *[Mm]odule*[cmd,symvers,order]
 
-# Generated rust-analyzer file
-rust-project.json
-
+# Virtual machine environment
+vm/
