+++ .gitmodules
@@ -2,7 +2,4 @@
 	path = linux
 	url = https://github.com/Rust-for-Linux/linux.git
 	ignore = all
-[submodule "busybox"]
-	path = busybox
-	url = https://github.com/mirror/busybox.git
-	ignore = all
+	shallow = true
