--- Kbuild
@@ -1,2 +1,0 @@
-obj-m := blackpill.o
-blackpill-y := $(RUST_OBJECTS)
