+++ README.md
@@ -10,15 +10,15 @@
 
 ![Rootkit simple architecture schema](assets/blackpill-rootkit-overview.drawio.png)
 
-todo
+- Hides itself from the modules list
 
 ## Development environment
 
 ### Description
 
 Multiple steps needs to be done before compiling our rootkit. The development environment is composed of :
-- a simple busybox initramfs providing essential tools
-- a custom compiled minimal kernel with Rust activated
+- a simple Alpine Linux image providing essential tools
+- a custom compiled kernel with Rust activated
 - a simple QEMU virtual machine accelerated by KVM
 
 Start by cloning the repository and its submodules :
@@ -26,59 +26,12 @@ Start by cloning the repository and its submodules :
 $ git clone git@github.com:DualHorizon/blackpill.git --recursive
 ```
 
-### QEMU
+### Important dependencies
 
 On an arch-based distribution :
 
 ```shell
-$ sudo pacman -S qemu-base qemu-desktop
-```
-
-### Busybox
-
-```shell
-$ cd busybox/
-$ make defconfig
-$ git am ../patches/busybox-ncurses-fix.patch
-$ make menuconfig
-# Then navigate into "Settings" and select "Build static binary (no shared libs)"
-```
-
-You can then compile and create the image :
-
-```shell
-$ make -j $(nproc)
-$ make install
-```
-
-If you have `TC` related errors, run the following command and recompile :
-
-```shell
-$ sed -i "s/CONFIG_TC=y/CONFIG_TC=n/g" .config
-```
-
-Once done, create the ramdisk image :
-
-```shell
-$ cd _install
-$ mkdir -p bin sbin etc proc sys dev usr/{s,}bin
-$ cp ../examples/inittab ./etc/
-$ sed "s/^tty.*//g" etc/inittab
-$ cat > "init" <<END
-#!/bin/sh
-
-mount -t proc none /proc
-mount -t sysfs none /sys
-mount -t devtmpfs none /dev
-
-ifconfig lo up
-ifconfig eth0 up
-udhcpc -i eth0
-
-/bin/sh
-END
-$ chmod +x init
-$ find . | cpio -H newc -o | gzip > ../ramdisk.img
+$ sudo pacman -S qemu-base qemu-desktop docker grub
 ```
 
 ### Linux kernel
@@ -99,26 +52,20 @@ $ cargo install --locked bindgen-cli
 Make sure everything is ok by running in folder `linux/` :
 
 ```shell
-$ cd linux
+$ pushd linux
 $ make LLVM=1 rustavailable
 Rust is available!
+$ popd
 ```
 
-Then apply the patch to create a new minimal kernel config (still in `linux/`):
+Launch the first time setup task :
 
 ```shell
-$ git am ../patches/qemu-busybox-min.patch
+$ make first-time-setup
 ```
 
-We can configure our kernel and build it :
-
-```shell
-$ make LLVM=1 allnoconfig qemu-busybox-min.config rust.config
-$ make LLVM=1 -j $(nproc)
-$ make LLVM=1 -j $(nproc) rust-analyzer
-$ cd ..
-$ make -C linux/ M=$PWD rust-analyzer
-```
+> [!IMPORTANT]
+> If you are asked of customizing options, press Enter each time.
 
 ### Rootkit
 
@@ -134,19 +81,17 @@ Launch the VM with :
 $ make vm
 ```
 
+Inside the VM, login with `root:root` (note: to be changed to automatic login) and enable the module :
+
+```shell
+$ modprobe blackpill
+# you can check kernel logs with
+$ dmesg
+```
+
 ## Usage
 
 todo
-
-## Roadmap
-
-**v0.1.0**
-
-- [ ] To-do
-
-**v0.2.0**
-
-- [ ] To-do
 
 ## Credits
