+++ rust-toolchain.toml
@@ -1,0 +1,5 @@
+[toolchain]
+channel = "nightly"
+components = [ "rustfmt", "llvm-tools-preview" ]
+targets = [ "x86_64-unknown-none" ]
+profile = "minimal"
