--- src/hypervisor/hardware_vt/mod.rs
@@ -1,2 +1,0 @@
-include! {"svm.rs"}
-include! {"vmx.rs"}
